Fix build resource paths

+++ Makefile
@@ -1,7 +1,9 @@
 submodule_path := SublimeText_Documentation
 local_path := $(submodule_path)/www.sublimetext.com
 local_index := $(local_path)/docs/index.html
-built_path := $(local_path)/sublime-text.docset
+docset := sublime-text.docset
+dashing_json := $(local_path)/dashing.json
+built_path := $(local_path)/$(docset)
 
 .PHONY: all
 all: clean pre-build build post-build
@@ -14,7 +16,7 @@ fix-html:
 	python fix_html.py
 
 build:
-	yq -j . dashing.yml > $(local_path)/dashing.json
+	yq -j . dashing.yml > $(dashing_json)
 	cd $(local_path) \
 	&& dashing build
 
@@ -25,9 +27,6 @@ post-build:
 
 .PHONY: clean
 clean:
+	[ -d "$(built_path)" ] && rm -r $(built_path) || true
+	[ -f "$(dashing_json)" ] && rm $(dashing_json) || true
 	git restore $(submodule_path) --recurse-submodules
-
-.PHONY: clean-more
-clean-more: clean
-	rm -r $(built_path)
-	rm $(local_path)/dashing.json
